pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-server-compare"
version = "0.1.0"
description = "A tool for comparing and synchronizing SQL Server database schemas"
authors = [
    { name = "Jose Freitas" }
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pyodbc>=4.0.39",
    "ttkbootstrap>=1.10.1",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
sqlcompare = "sql_compare.main:main"

[tool.black]
line-length = 100
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
